% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claims}
\alias{claims}
\title{A sample of 100,000 medical insurance claims}
\format{
A data frame with 100,000 rows and 56 variables:
\describe{
  \item{form_year}{Year of claim}
  \item{age}{Age of patient}
  ...
}
}
\source{
\url{}
}
\usage{
claims
}
\description{
A sample of 100,000 medical insurance claims
}
\keyword{datasets}
